refactor: fix hotels components
