+++ src/components/SelectHotels/index.js
@@ -1,0 +1,5 @@
+import ValidateHotel from './ValidateHotels';
+
+export default function SelectHotels() {
+  return <ValidateHotel />;
+}
